requirements.txt

python-dotenv
fastapi
uvicorn[standard]
pydantic
httpx[socks]
git+https://github.com/jvkabum/hcaptcha-challenger.git@main#egg=hcaptcha-challenger
camoufox
browserforge